+++ resources/certificate.rb
@@ -32,11 +32,7 @@ action :install do
   java_home = new_resource.java_home
   keytool = "#{java_home}/bin/keytool"
   truststore = if new_resource.keystore_path.empty?
-                 if ::File.exist?("#{java_home}/lib/security/cacerts")
-                   "#{java_home}/lib/security/cacerts"
-                 else
-                   "#{java_home}/jre/lib/security/cacerts"
-                 end
+                 truststore_default_location
                else
                  new_resource.keystore_path
                end
@@ -92,11 +88,7 @@ end
 action :remove do
   certalias = new_resource.name
   truststore = if new_resource.keystore_path.nil?
-                 if ::File.exist?("#{java_home}/jre/lib/security/cacerts")
-                   "#{java_home}/jre/lib/security/cacerts"
-                 else
-                   "#{java_home}/lib/security/cacerts"
-                 end
+                 truststore_default_location
                else
                  new_resource.keystore_path
                end
@@ -123,24 +115,30 @@ action :remove do
 end
 
 action_class do
-  require 'openssl'
-
   def fetch_certdata
     return IO.read(new_resource.cert_file) unless new_resource.cert_file.nil?
 
     certendpoint = new_resource.ssl_endpoint
     unless certendpoint.nil?
-      ctx = OpenSSL::SSL::SSLContext.new
-      dest = certendpoint.split(/:/)
-      sock = TCPSocket.new(dest[0], dest.length >= 2 ? dest[1] : 443)
-      ssl = OpenSSL::SSL::SSLSocket.new(sock, ctx)
-      # ssl.hostname = dest[0] # for SNI
-      ssl.connect
-      cert = ssl.peer_cert
-      return cert.to_pem unless cert.nil?
+      cmd = Mixlib::ShellOut.new("echo QUIT | openssl s_client -showcerts -connect #{certendpoint} 2> /dev/null | openssl x509")
+      cmd.run_command
+      Chef::Log.debug(cmd.format_for_exception)
+
+      Chef::Application.fatal!("Error returned when attempting to retrieve certificate from remote endpoint #{certendpoint}: #{cmd.exitstatus}", cmd.exitstatus) unless cmd.exitstatus == 0
+
+      certout = cmd.stdout
+      return certout unless certout.empty?
       Chef::Application.fatal!("Unable to parse certificate from openssl query of #{certendpoint}.", 999)
     end
 
     Chef::Application.fatal!('At least one of cert_data, cert_file or ssl_endpoint attributes must be provided.', 999)
+  end
+
+  def truststore_default_location
+    if node['java']['jdk_version'].to_i > 8
+      "#{node['java']['java_home']}/lib/security/cacerts"
+    else
+      "#{node['java']['java_home']}/jre/lib/security/cacerts"
+    end
   end
 end